Fix opacity issue with caption container
When changing the caption containers opacity this will directly impact the opacity of the content as well, which may not be desired behavior (in my case). To allow for independent changes to the background color opacity two more elements are required as usually with modal dialogs. This PR adds these elements - one for the backdrop and one for the caption text allowing to access and style it.